Encapsulate different AI into different work item
In order to allow the whole program to run asynchronously, we need to
encapsulate different AI function into different work item so we can
place them in CMWQ.

Now the implementation has two kinds of tasklet, "ai_game_tasklet" and
"simrupt_tasklet", the former is used to schedule tic-tac-toe game
between AI algorithm, the latter is responsible of drawing the chess
board layout and send it to the userspace.

Future improvements are needed to make sure we only draw the new status
of the chess board rather than draw the same status of chess board
repeatedly.