Implement a new AI algorithm and fix atomic bug
Implement another AI algorithm which is negamax, also provide it's
helper package zobrist. However, we've shrink the size of
HASH_TABLE_SIZE or it'll be too large to be kmalloc().

The atomic variable "turn" was a wrong implementation which will cause
numerous error message in kernel space, so we change "turn" to the type
of "char" and use memory barrier and the macros "READ_ONCE()",
"WRITE_ONCE()" to provide the data consistency between different
processor.