Ensure the execution order of AI algorithms
In the previous versions of implementation, there might be numerous AI
algorithms' work item present in the CMWQ, although we can ensure the
enqueue order of different work item, we can't ensure the execution
order is correct due to lack of ability to predict the execution time.

We add a variable "finish" to determine whether the current AI algorithm
is still under exeuction context, if yes then we can't enqueue a new AI
algorithm into the queue until it's done.

Doing so can ensure both the enqueue order and execution order of
different AI algorithm, without the need to check for permission to
execute inside the work function.