Change name of tasklet handler
Originally, the prefix of tasklet handler is "simrupt", now the module
isn't named as "simrupt" anymore. We utilize the fact that the tasklet
handler is indeed a schduler which enqueue different work items into
CMWQ during tic-tac-toe games, so we call it "game_tasklet" instead.

Also the work function "simrupt_work_func" is actually doing the work of
drawing the chess board to a buffer and then copy the data from the
buffer into kfifo buffer so we can send the chess board to user space,
that's the reason we call it "drawboard_work_func" now.