Fix bug when resume to display game
Last version of implementation can't resume to display the chess board
after we stop to display it. We introduce I/O multiplexing mechanism in
order to listen to both STDIN_FILENO and device driver attribute file.
This way we can continue to listen to keyboard event even when the
kernel stops sendding data to user space.