Implement xoroshiro random number generator
Implementation of xoroshiro, another kind of PRNG. About the detailed
concept of the algorithm please refer to the paper "Scrambled Linear
Pseudorandom Number Generators".

We implement this algorithm in order to replace the wyhash algorithm
used in MCTS simulation step. Note that we're using the ++scrambler in
the xoro_next() function.