Calculate load average of AI algorithm
Introduce the calculation of load average for mcts algorithm, taking the
active nodes in the current mcts exeuction context as calculation units.
The function "mcts_calc_load()" takes reference from the method of
counting cpu load average in linux kernel, the concept is EWMA.

Utilize EWMA to calculate the load average of MCTS algorithm so we can
know how much recources the algorithm context is consuming.