Fix incorrect open count check in release function
In the original code, the kernel module stops producing new
boards when atomic_dec_and_test(&open_cnt) returns zero, but
it will return zero when the resulting value of open_cnt is
nonzero. Therefore, in the origional code, if we kill the
user program who is interacting with this kernel module, it
won't stop producing new boards, but it is supposed to stop
producing new boards when open_cnt becomes zero.

As a result, the code is modified so that it stops producing
new boards when atomic_dec_and_test(&open_cnt) returns one.